fix(mas): split egui keymap helpers out of feature-gated hotkeys module
